完善反射测试案例

+++ src/test/java/com/edison/testJunit/oth/ii_2_XML/parse_big_xml/base/XmlBase.java
@@ -210,7 +210,7 @@ public  class XmlBase {
 						setMethod.invoke(this, list);
 				    }
 				} else if (fieldType == String.class || fieldType == Integer.class || fieldType == Long.class||fieldType==Boolean.class
-					|| fieldType == Float.class || fieldType == Double.class ||fieldType == CHAR.class||fieldType == Long.class
+					|| fieldType == Float.class || fieldType == Double.class ||fieldType == CHAR.class||fieldType == Byte.class
 					||fieldType==int.class||fieldType==byte.class||fieldType==char.class||fieldType==short.class
 					||fieldType==boolean.class||fieldType==double.class||fieldType==float.class||fieldType==long.class) {
 					//TODO 基本类型及其包装类型的判断以及赋值最好提取出来
